move output enum to config types
